requirements.txt
asgiref==3.4.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-polymorphic==3.0.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.0
djangorestframework-recursive==0.1.2
djangorestframework-serializer-extensions==2.0.1
hashids==1.3.1
inflection==0.5.1
Markdown==3.3.4
psycopg2-binary==2.9.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
